requirements.txt
absl-py
jupyter
rich
scikit-learn
seaborn
torch
tqdm
